+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 3.3.1
+
+Fixed
+
+- Bug where the penultimate day in a Pill Schedule with an "X Days On, X Days Off" expiration interval would 
+    incorrectly set the next date after taking.
 ## 3.3.0
 
 Changed